Separate /jitpack subcommands
