+++ doxxy-bot/src/main/kotlin/dev/freya02/doxxy/bot/versioning/jitpack/JitpackPrService.kt
@@ -2,20 +2,23 @@ package dev.freya02.doxxy.bot.versioning.jitpack
 
 import dev.freya02.doxxy.bot.config.PullUpdaterConfig
 import dev.freya02.doxxy.bot.versioning.LibraryType
-import dev.freya02.doxxy.bot.versioning.github.PullRequest
-import dev.freya02.doxxy.bot.versioning.github.PullRequestCache
-import dev.freya02.doxxy.bot.versioning.github.UpdatedBranch
+import dev.freya02.doxxy.bot.versioning.github.*
 import dev.freya02.doxxy.bot.versioning.jitpack.pullupdater.PullUpdateException
 import dev.freya02.doxxy.bot.versioning.jitpack.pullupdater.PullUpdater
 import dev.minn.jda.ktx.messages.send
 import io.github.freya022.botcommands.api.core.service.annotations.BService
 import io.github.oshai.kotlinlogging.KotlinLogging
+import kotlinx.coroutines.async
+import kotlinx.coroutines.coroutineScope
 import net.dv8tion.jda.api.interactions.InteractionHook
 
 private val logger = KotlinLogging.logger { }
 
 @BService
-class JitpackPrService(private val pullUpdaterConfig: PullUpdaterConfig) {
+class JitpackPrService(
+    private val pullUpdaterConfig: PullUpdaterConfig,
+    private val client: GithubClient,
+) {
     private val bcPullRequestCache = PullRequestCache(LibraryType.BOT_COMMANDS, null)
     private val jdaPullRequestCache = PullRequestCache(LibraryType.JDA, "master")
     private val jdaKtxPullRequestCache = PullRequestCache(LibraryType.JDA_KTX, "master")
@@ -56,4 +59,35 @@ class JitpackPrService(private val pullUpdaterConfig: PullUpdaterConfig) {
     fun canUsePullUpdate(libraryType: LibraryType): Boolean {
         return libraryType == LibraryType.JDA && pullUpdaterConfig.enable
     }
+
+    suspend fun getAdditionalPRDetails(libraryType: LibraryType, pullRequest: PullRequest): AdditionalPullRequestDetails {
+        return coroutineScope {
+            val updatedPR = async {
+                client.getPullRequest(libraryType.githubOwnerName, libraryType.githubRepoName, pullRequest.number)
+            }
+
+            val commitComparisons = async {
+                val base = pullRequest.base.label
+                val head = pullRequest.head.label
+
+                client.compareCommits(libraryType.githubOwnerName, libraryType.githubRepoName, base, head)
+                    .toUpdatedCommitComparisons()
+            }
+
+            val reverseCommitComparisons = async {
+                val base = pullRequest.base.label
+                val head = pullRequest.head.label
+
+                client.compareCommits(libraryType.githubOwnerName, libraryType.githubRepoName, head, base)
+            }
+
+            AdditionalPullRequestDetails(updatedPR.await(), commitComparisons.await(), reverseCommitComparisons.await())
+        }
+    }
+
+    data class AdditionalPullRequestDetails(
+        val updatedPR: PullRequest,
+        val commitComparisons: UpdatedCommitComparisons,
+        val reverseCommitComparisons: CommitComparisons,
+    )
 }
